refactor into separate like entity
