+++ backend/src/entities/likes/route.ts
@@ -1,0 +1,18 @@
+import { PostgresJsDatabase } from "drizzle-orm/postgres-js";
+import { Hono } from "hono";
+import { LikeTransaction, LikeTransactionImpl } from "./transaction";
+import { LikeService, LikeServiceImpl } from "./service";
+import { LikeController, LikeControllerImpl } from "./controller";
+
+export const likeRoutes = (db: PostgresJsDatabase): Hono => {
+  const like = new Hono();
+
+  const likeTransaction: LikeTransaction = new LikeTransactionImpl(db);
+  const likeService: LikeService = new LikeServiceImpl(likeTransaction);
+  const postController: LikeController = new LikeControllerImpl(likeService);
+
+  like.patch("/", (ctx) => postController.toggleLike(ctx));
+  like.get("/", (ctx) => postController.getLikeUsers(ctx));
+
+  return like;
+};
