+++ backend/src/entities/posts/controller.ts
@@ -9,16 +9,13 @@ import {
   updatePostValidate,
 } from "./validator";
 import { Status } from "../../constants/http";
-import { DEL_POST, LIKE_USERS, POST_API, TOGGLE_LIKE } from "../../types/api/routes/posts";
-import { paginationSchema } from "../../utilities/pagination";
+import { DEL_POST, POST_API } from "../../types/api/routes/posts";
 
 export interface PostController {
   createPost(ctx: Context): Promise<POST_API>;
   getPost(ctx: Context): Promise<POST_API>;
   updatePost(ctx: Context): Promise<POST_API>;
   deletePost(ctx: Context): Promise<DEL_POST>;
-  toggleLike(ctx: Context): Promise<TOGGLE_LIKE>;
-  getLikeUsers(ctx: Context): Promise<LIKE_USERS>;
 }
 
 export class PostControllerImpl implements PostController {
@@ -92,28 +89,5 @@ export class PostControllerImpl implements PostController {
       return ctx.json({ message: "Successfully delete post" }, Status.OK);
     };
     return await handleAppError(deletePostImpl)(ctx);
-  }
-
-  async toggleLike(ctx: Context): Promise<TOGGLE_LIKE> {
-    const toggleLikeImpl = async () => {
-      const postId = parseUUID(ctx.req.param("id"));
-      const userId = ctx.get("userId");
-      const isLiking = await this.postService.toggleLike({ id: postId, userId });
-      const action = isLiking ? "like" : "unlike";
-      return ctx.json({ message: `Successfully ${action} post` }, Status.OK);
-    };
-    return await handleAppError(toggleLikeImpl)(ctx);
-  }
-
-  async getLikeUsers(ctx: Context): Promise<LIKE_USERS> {
-    const getLikeUsersImpl = async () => {
-      const { limit, page } = ctx.req.query();
-      const postId = parseUUID(ctx.req.param("id"));
-      const userId = ctx.get("userId");
-      const pagination = paginationSchema.parse({ limit, page });
-      const users = await this.postService.getLikeUsers({ id: postId, userId, ...pagination });
-      return ctx.json(users, Status.OK);
-    };
-    return await handleAppError(getLikeUsersImpl)(ctx);
   }
 }
