+++ backend/src/routes/init.ts
@@ -6,6 +6,7 @@ import { getOpenAPISpecification } from "../utilities/docs";
 import { HEALTHCHECK } from "../types/api/routes/healthcheck";
 import { groupRoutes } from "../entities/groups/route";
 import { postRoutes } from "../entities/posts/route";
+import { likeRoutes } from "../entities/likes/route";
 
 export const setUpRoutes = (app: Hono, db: PostgresJsDatabase) => {
   // api documentation
@@ -37,6 +38,7 @@ const apiRoutes = (db: PostgresJsDatabase): Hono => {
   api.route("/users", userRoutes(db));
   api.route("/groups", groupRoutes(db));
   api.route("/", postRoutes(db));
+  api.route("/posts/:id/likes", likeRoutes(db));
 
   return api;
 };
